# talent-match-system/docs/api/api-contract.yaml
openapi: 3.0.0
info:
  title: 人才职位智能匹配与能力图谱系统 API
  version: 1.0.0
  description: 基于AI的人才职位智能匹配与能力图谱系统API接口文档
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
paths:
...
  /api/job:
    post:
      summary: 创建岗位
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                job_name:
                  type: string
                  description: 岗位名称
                company_name:
                  type: string
                  description: 公司名称
                company_size:
                  type: string
                  description: 公司规模
                company_industry:
                  type: string
                  description: 所属行业
                company_intro:
                  type: string
                  description: 公司介绍
                location:
                  type: string
                  description: 工作地点
                education:
                  type: string
                  description: 学历要求
                experience:
                  type: string
                  description: 经验要求
                salary_min:
                  type: integer
                  description: 最低薪资
                salary_max:
                  type: integer
                  description: 最高薪资
                job_desc:
                  type: string
                  description: 岗位职责
                requirements:
                  type: string
                  description: 任职要求
                skills:
                  type: array
                  items:
                    type: string
                  description: 技能要求
      responses:
        '200':
          description: 创建成功

    get:
      summary: 获取岗位列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
  /api/job/{job_id}:
    get:
      summary: 获取岗位详情
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
        '404':
          description: 岗位不存在

    put:
      summary: 更新岗位
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                job_name:
                  type: string
                location:
                  type: string
                education:
                  type: string
                experience:
                  type: string
                salary_min:
                  type: integer
                salary_max:
                  type: integer
                job_desc:
                  type: string
                requirements:
                  type: string
      responses:
        '200':
          description: 更新成功
        '404':
          description: 岗位不存在

    delete:
      summary: 删除岗位
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
        '404':
          description: 岗位不存在
  /api/job/upload:
    post:
      summary: 上传岗位要求文件
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                job_name:
                  type: string
                company_name:
                  type: string
                company_size:
                  type: string
                company_industry:
                  type: string
                company_intro:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 上传成功
  /api/job/extract-job:
    post:
      summary: 解析岗位信息
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: 解析成功
  /api/job/ability/create:
    post:
      summary: 创建岗位能力图谱
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                job_id:
                  type: integer
                  description: 岗位ID
                abilities:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      weight:
                        type: number
                      level:
                        type: integer
      responses:
        '200':
          description: 创建成功
  /api/job/ability/tree/{job_id}:
    get:
      summary: 获取岗位能力树
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
  /api/job/ability/update/{node_id}:
    put:
      summary: 更新岗位能力节点
      security:
        - bearerAuth: []
      parameters:
        - name: node_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 更新成功
  /api/job/ability/delete/{job_id}:
    delete:
      summary: 删除岗位能力图谱
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
  /api/match:
    post:
      summary: 简历岗位匹配
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                resume_id:
                  type: integer
                  description: 简历ID
                job_id:
                  type: integer
                  description: 岗位ID
      responses:
        '200':
          description: 匹配成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  match_id:
                    type: integer
                  score:
                    type: number
                    description: 匹配分数
                  details:
                    type: object
                    description: 匹配详情
  /api/match/records:
    get:
      summary: 获取匹配记录列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    resume_id:
                      type: integer
                    job_id:
                      type: integer
                    score:
                      type: number
                    status:
                      type: string
                    created_at:
                      type: string
                      format: date-time
  /api/match/recommendations:
    get:
      summary: 获取推荐岗位/简历
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
  /api/match/{match_id}:
    get:
      summary: 获取匹配详情
      security:
        - bearerAuth: []
      parameters:
        - name: match_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
        '404':
          description: 匹配记录不存在
...
  /api/ability/tree/{user_id}:
    get:
      summary: 获取用户能力树
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: resume_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
...
  /api/ability/create:
    post:
      summary: 创建能力图谱
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                ability_data:
                  type: object
      responses:
        '200':
          description: 创建成功
  /api/ability/update/{node_id}:
    put:
      summary: 更新能力节点
      parameters:
        - name: node_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 更新成功
  /api/ability/delete/{user_id}:
    delete:
      summary: 删除能力图谱
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT